+++ src/client/allEvents.ts
@@ -61,3 +61,74 @@ const getEventPage = async (
             })
     );
 };
+
+export const getAllEventsAfterEvent = async (
+    tenantId: string,
+    topic: string,
+    types: string[],
+    eventId: string,
+    perPage: number,
+    handler: HandlerFunc,
+    serviceClient: ServiceClient
+): Promise<void> => {
+    let page = 0;
+
+    while (true) {
+        const events = await getEventPageAfterEvent(
+            tenantId,
+            topic,
+            types,
+            eventId,
+            perPage,
+            page,
+            serviceClient
+        );
+
+        if (events.length === 0) {
+            return;
+        }
+
+        for (const eventData of events) {
+            const event = getSubscriptionEvent(eventData);
+            if (event) {
+                await handler(event);
+            }
+        }
+
+        page++;
+    }
+};
+
+const getEventPageAfterEvent = async (
+    tenantId: string,
+    topic: string,
+    types: string[],
+    eventId: string,
+    perPage: number,
+    page: number,
+    serviceClient: ServiceClient
+) => {
+    return retry(
+        () =>
+            new Promise<Event[]>((resolve, reject) => {
+                serviceClient.paginateEventsAfterEventId(
+                    {
+                        tenantId,
+                        topic,
+                        types,
+                        eventId,
+                        page: page.toString(),
+                        perPage: perPage.toString(),
+                    },
+                    async (error, data) => {
+                        if (error) {
+                            reject(error);
+                            return;
+                        }
+
+                        resolve(data.events);
+                    }
+                );
+            })
+    );
+};